+++ src/FishyFlip/ATProtoModeration.cs
@@ -24,6 +24,15 @@ public sealed class ATProtoModeration
 
     private HttpClient Client => this.proto.Client;
 
+    /// <summary>
+    /// Creates a moderation report for a post asynchronously.
+    /// </summary>
+    /// <param name="reasonType">The type of the moderation reason.</param>
+    /// <param name="uri">The URI of the post.</param>
+    /// <param name="cid">The CID of the post.</param>
+    /// <param name="reason">The reason for the moderation report. This is optional.</param>
+    /// <param name="cancellationToken">A token that may be used to cancel the operation. This is optional.</param>
+    /// <returns>A task that represents the asynchronous operation. The task result contains a <see cref="Result{ModerationRecord}"/> that represents the moderation record.</returns>
     public Task<Result<ModerationRecord?>> CreateModerationReportPostAsync(ModerationReasonType reasonType, ATUri uri, Cid cid, string? reason = default, CancellationToken cancellationToken = default)
     {
         return this.Client.Post<CreateModerationReportPost, ModerationRecord?>(
@@ -36,6 +45,14 @@ public sealed class ATProtoModeration
                        this.Options.Logger);
     }
 
+    /// <summary>
+    /// Creates a moderation report for a repository asynchronously.
+    /// </summary>
+    /// <param name="reasonType">The type of the moderation reason.</param>
+    /// <param name="subject">The decentralized identifier (DID) of the subject.</param>
+    /// <param name="reason">The reason for the moderation report. This is optional.</param>
+    /// <param name="cancellationToken">A token that may be used to cancel the operation. This is optional.</param>
+    /// <returns>A task that represents the asynchronous operation. The task result contains a <see cref="Result{ModerationRecord}"/> that represents the moderation record.</returns>
     public Task<Result<ModerationRecord?>> CreateModerationReportRepoAsync(ModerationReasonType reasonType, ATDid subject, string? reason = default, CancellationToken cancellationToken = default)
     {
         return this.Client.Post<CreateModerationReportRepo, ModerationRecord?>(
