+++ src/FishyFlip/FishyFlip.csproj
@@ -5,9 +5,10 @@
 		<ImplicitUsings>enable</ImplicitUsings>
 		<Nullable>enable</Nullable>
 		<IsPackable>true</IsPackable>
-		<TreatWarningsAsErrors>false</TreatWarningsAsErrors>
+		<TreatWarningsAsErrors>true</TreatWarningsAsErrors>
 		<WarningsAsErrors>nullable</WarningsAsErrors>
 		<IsAotCompatible>true</IsAotCompatible>
+		<NoWarn>$(NoWarn);SA0001</NoWarn>
 	</PropertyGroup>
 	<ItemGroup>
 		<PackageReference Include="Microsoft.Extensions.Logging.Abstractions" />
