+++ src/FishyFlip/Models/FeedPostList.cs
@@ -4,4 +4,7 @@
 
 namespace FishyFlip.Models;
 
+/// <summary>
+/// Represents a list of feed posts.
+/// </summary>
 public record FeedPostList(FeedViewPost[] Feed, string? Cursor);
