+++ src/FishyFlip/Models/Internal/CreateMuteListRecord.cs
@@ -1,7 +1,10 @@
-// <copyright file="CreateMuteRecord.cs" company="Drastic Actions">
+// <copyright file="CreateMuteListRecord.cs" company="Drastic Actions">
 // Copyright (c) Drastic Actions. All rights reserved.
 // </copyright>
 
 namespace FishyFlip.Models.Internal;
 
+/// <summary>
+/// Represents a record for creating a mute list.
+/// </summary>
 public record CreateMuteListRecord(ATUri List);
