+++ src/FishyFlip/Models/ListViewRecord.cs
@@ -4,4 +4,7 @@
 
 namespace FishyFlip.Models;
 
+/// <summary>
+/// Represents a record containing an array of ListView objects and a cursor.
+/// </summary>
 public record ListViewRecord(ListView[] Lists, string? Cursor);
