+++ src/FishyFlip/Models/Service.cs
@@ -1,4 +1,4 @@
-// <copyright file="DescribeRepo.cs" company="Drastic Actions">
+// <copyright file="Service.cs" company="Drastic Actions">
 // Copyright (c) Drastic Actions. All rights reserved.
 // </copyright>
