+++ src/FishyFlip/Tools/Json/AtHandlerJsonConverter.cs
@@ -4,8 +4,12 @@
 
 namespace FishyFlip.Tools.Json;
 
+/// <summary>
+/// Provides a converter for the ATHandle type for JSON serialization and deserialization.
+/// </summary>
 public class AtHandlerJsonConverter : JsonConverter<ATHandle>
 {
+    /// <inheritdoc/>
     public override ATHandle? Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
     {
         string? value = reader.GetString();
@@ -25,6 +29,7 @@ public class AtHandlerJsonConverter : JsonConverter<ATHandle>
         }
     }
 
+    /// <inheritdoc/>
     public override void Write(Utf8JsonWriter writer, ATHandle? value, JsonSerializerOptions options)
     {
         writer.WriteStringValue(value?.ToString());
